refactor: exception 패캐지를 application 외부로 분리
