+++ src/main/java/com/ziine/global/error/presentation/GlobalExceptionHandler.java
@@ -1,9 +1,9 @@
 package com.ziine.global.error.presentation;
 
-import com.ziine.admin.auth.application.exception.AdminUnauthorizedException;
-import com.ziine.global.error.application.exception.BusinessException;
+import com.ziine.admin.auth.exception.AdminUnauthorizedException;
 import com.ziine.global.error.domain.ErrorCode;
 import com.ziine.global.error.dto.ErrorResponseDto;
+import com.ziine.global.error.exception.BusinessException;
 import jakarta.annotation.Nullable;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.stream.Collectors;
